+++ pdoc/search.py
@@ -39,7 +39,7 @@ If you wish to hide the search box, you can add
 {% block search %}{% endblock %}
 {% block search_js %}{% endblock %}
 ```
-in your [`module.html.jinja2` template](../pdoc.html#editing-pdocs-html-template).
+in your [`module.html.jinja2` template](../pdoc.html#edit-pdocs-html-template).
 """
 from __future__ import annotations